everything is working except create a note
