+++ models/Note.js
@@ -1,5 +1,4 @@
 var mongoose = require("mongoose");
-
 var Schema = mongoose.Schema;
 
 var NoteSchema = new Schema({
