+++ routes/index.js
@@ -1,0 +1,8 @@
+var router = require("express").Router();
+var apiRouter = require("./apiRoutes");
+var htmlRouter = require("./htmlRoutes");
+
+router.use("/articles", apiRouter);
+router.use("/", htmlRouter);
+
+module.exports = router;
